Release new paths UI for everyone
