+++ frontend/src/scenes/funnels/FunnelStepDropdown.tsx
@@ -5,20 +5,13 @@ import { FunnelPathType, PathType, InsightType, AvailableFeature } from '~/types
 import { funnelLogic } from './funnelLogic'
 import { useValues } from 'kea'
 import { insightLogic } from 'scenes/insights/insightLogic'
-import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
-import { FEATURE_FLAGS } from 'lib/constants'
 import { userLogic } from 'scenes/userLogic'
 
 export function FunnelStepDropdown({ index }: { index: number }): JSX.Element | null {
     const { insightProps } = useValues(insightLogic)
     const logic = funnelLogic(insightProps)
     const { propertiesForUrl: filterProps, filters } = useValues(logic)
-    const { featureFlags } = useValues(featureFlagLogic)
     const { user } = useValues(userLogic)
-
-    if (!featureFlags[FEATURE_FLAGS.NEW_PATHS_UI]) {
-        return null
-    }
 
     if (!user?.organization?.available_features?.includes(AvailableFeature.PATHS_ADVANCED)) {
         // TODO: Consider showing the options but disabled with a prompt to upgrade
