+++ frontend/src/scenes/insights/InsightTabs/PathTab/PathTab.scss
